+++ tts_app/voice_api/views.py
@@ -566,7 +566,7 @@ def voice_gpt_sovits_api():
                                 config.gpt_sovits_config.presets.get("default").prompt_lang, str)
         top_k = get_param(request_data, "top_k", config.gpt_sovits_config.top_k, int)
         top_p = get_param(request_data, "top_p", config.gpt_sovits_config.top_p, float)
-        temperature = get_param(request_data, "temperature", config.gpt_sovits_config.temperature, int)
+        temperature = get_param(request_data, "temperature", config.gpt_sovits_config.temperature, float)
         # use_streaming = get_param(request_data, 'streaming', config.gpt_sovits_config.use_streaming, bool)
     except Exception as e:
         logger.error(f"[{ModelType.GPT_SOVITS.value}] {e}")
